Prevent Rollup cycle warning by moving mergeOptions to utilities.
