+++ src/utilities/index.ts
@@ -92,5 +92,6 @@ export * from './common/canUse';
 export * from './common/compact';
 export * from './common/makeUniqueId';
 export * from './common/stringifyForDisplay';
+export * from './common/mergeOptions';
 
 export * from './types/IsStrictlyAny';
